add exception handling

+++ src/main/java/com/takehomeassignment/tasks/controller/TaskController.java
@@ -2,7 +2,10 @@ package com.takehomeassignment.tasks.controller;
 
 import java.util.List;
 
+import com.takehomeassignment.tasks.exceptions.BadInputException;
+import com.takehomeassignment.tasks.exceptions.NoSuchElementFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -20,6 +23,10 @@ public class TaskController {
 
     @RequestMapping(value="/tasks", method=RequestMethod.POST)
     public Task createTask(@RequestBody Task t) {
+        if (t.getTitle().isBlank()) {
+            throw new BadInputException("Title is required");
+        }
+
         return taskService.createTask(t);
     }
 
@@ -30,16 +37,34 @@ public class TaskController {
 
     @RequestMapping(value="/tasks/{taskId}", method=RequestMethod.GET)
     public Task getTaskById(@PathVariable(value = "taskId") Long id) {
+        if (!taskService.isExistingTask(id)) {
+            throw new NoSuchElementFoundException("Task ID " + id + " does not exist");
+        }
+
         return taskService.getTaskById(id);
     }
 
     @RequestMapping(value="/tasks/{taskId}", method=RequestMethod.PUT)
-    public Task updateTask(@PathVariable(value = "taskId") Long id, @RequestBody Task taskDetails) {
-        return taskService.updateTask(id, taskDetails);
+    public Task updateTask(@PathVariable(value = "taskId") Long id, @RequestBody Task updatedTask) {
+        if (!taskService.isExistingTask(id)) {
+            throw new NoSuchElementFoundException("Task ID " + id + " does not exist");
+        }
+
+        if (updatedTask.getTitle() != null && updatedTask.getTitle().isEmpty()) {
+            throw new BadInputException("Title cannot be empty");
+        }
+
+        return taskService.updateTask(id, updatedTask);
     }
 
     @RequestMapping(value="/tasks/{taskId}", method=RequestMethod.DELETE)
-    public void deleteEmployees(@PathVariable(value = "taskId") Long id) {
+    public ResponseEntity<String> deleteEmployees(@PathVariable(value = "taskId") Long id) {
+        if (!taskService.isExistingTask(id)) {
+            throw new NoSuchElementFoundException("Task ID " + id + " does not exist");
+        }
+
         taskService.deleteTask(id);
+
+        return ResponseEntity.ok().body("Successfully deleted task " + id);
     }
 }
